layout/forms/test/test_bug620936.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=620936
-->
<head>
  <title>Test for Bug 620936</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=620936">Mozilla Bug 620936</a>
<p id="display"></p>
<div id="content">
  <input value="foo">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 620936 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var i = document.querySelector("input");
  i.focus();
  i.setSelectionRange(100, 100);
  is(i.selectionStart, 3, "The selection should be set to the end of the text");
  is(i.selectionEnd, 3, "The selection should be set to the end of the text");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
